+++ tests/testthat/test_query.R
@@ -71,6 +71,18 @@ test_that("Can upload via Google Cloud Storage",{
                   schema = user_schema)
 })
 
+test_that("Can upload nested JSON",{
+  skip_on_cran()
+  skip_if_no_env_auth("BQ_AUTH_FILE")
+  
+  the_list <- list(list(col1 = "yes", col2 = "no", col3 = list(nest1 = 1, nest2 = 3), col4 = "oh"),
+                   list(col1 = "yes2", col2 = "n2o", col3 = list(nest1 = 5, nest2 = 7), col4 = "oh2"), 
+                   list(col1 = "yes3", col2 = "no3", col3 = list(nest1 = 7, nest2 = 55), col4 = "oh3"))
+  bqr_upload_data(datasetId = "test", 
+                  tableId = "nested_list_json", 
+                  upload_data = the_list, autodetect = TRUE)
+})
+
 context("List tables")
 
 test_that("Can list tables", {